refactor(ratos): move BackgroundDisplayStatusProgressHandler class to ratos.py for general reuse
